// src/client/models/UserType.ts
export interface UserType {
  activityRate?: number;
  nest: { items: [number] };
  collection?: string[];
  description?: string;
  firstName?: string;
  birthdate?: string;
  lastName?: string;
  imgSrc?: string;
  title?: string;
  email?: string;
  id: string;
  username: string;
  password: string;
}